Quick cleanup to remove eslint disable rule
